refactor: HomeItem 이름 변경 및 DesignSystem 모듈로 이동 (#36)
